chore: move FLUSHSLOTS to  as a subcommand
